requirements.txt
streamlit==1.40.1
python-dotenv==1.0.1

langchain==1.2.0
langchain-groq==1.1.1
langchain-mcp-adapters==0.2.1
langgraph==1.0.5